+++ order.hpp
@@ -4,13 +4,13 @@
 struct ClientOrder 
 {
     uint32_t price;
-    uint32_t quantity;
+    uint16_t quantity;
     uint16_t orderID;
     uint8_t type; // 0 for buy, 1 for sell, 2 for cancel, 3 for market buy, 4 for market sell
 };
 
 struct EngineOrder {
-    uint32_t quantity;
+    uint16_t quantity;
     uint16_t orderID;
     uint8_t  levelNext;
     uint8_t  levelPrev;
@@ -27,7 +27,7 @@ struct GlobalOrderInfo
 struct Event
 {
     uint32_t price;
-    uint32_t quantity;
+    uint16_t quantity;
     uint16_t orderID;
     uint8_t type; // 0 for cancel, 1 for fill
     uint8_t side; // 0 for buy, 1 for sell
